Player Area Block Saved
Logic is now, when you complete the two tasks, it will update its location, and if it was not loaded from a save, It will quickly save the game, so it saves its new location. Once it goes it its new location, it will destroy itself, even when loaded into the game!